+++ src/LDP.ts
@@ -1,13 +1,15 @@
-import * as AccessPoint from "./LDP/AccessPoint";
 import * as BasicContainer from "./LDP/BasicContainer";
 import * as Container from "./LDP/Container";
+import * as DirectContainer from "./LDP/DirectContainer";
+import * as IndirectContainer from "./LDP/IndirectContainer";
 import * as PersistedContainer from "./LDP/PersistedContainer";
 import * as RDFSource from "./LDP/RDFSource";
 
 export {
-	AccessPoint,
 	BasicContainer,
 	Container,
+	DirectContainer,
+	IndirectContainer,
 	PersistedContainer,
 	RDFSource,
 };