+++ src/LDP/Container.spec.ts
@@ -134,61 +134,74 @@ describe( module( "Carbon/LDP/Container" ), ():void => {
 				pointer = Document.Factory.create();
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( false );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( false );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.RDFSource );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( false );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.RDFSource );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( false );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.BasicContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.DirectContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.IndirectContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.BasicContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.DirectContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.IndirectContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.RDFSource );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.BasicContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.RDFSource );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.DirectContainer );
 				expect( Container.Factory.hasRDFClass( pointer ) ).toBe( true );
 
-				pointer = Document.Factory.create( "http://example.com/resource/" );
+				pointer = Document.Factory.create();
+				pointer.id = "http://example.com/resource/";
 				(<Document.Class> pointer).types.push( NS.LDP.Class.RDFSource );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.Container );
 				(<Document.Class> pointer).types.push( NS.LDP.Class.IndirectContainer );